Manage params[:move] consistently across "staged" controllers
In order:

1. if 'move:next' is present, set it to 'next'
2. if 'move:back' is present, (but move:next isn't) set it to 'back'
3. use 'move' value if present
4. after that if it's not 'next' or 'back' default it to 'next' (only forwards!)

We also change from using ``<button name="move" value="<next-or-back>">`` to ``<input type="submit" name="move:<next-or-back>"/>``.  This allows us to work around 2 bugs in IE6 and IE7:

a) they submit the value of all buttons, not just the one clicked
b) for the value of the param value they use the ``innerHTML`` of the ``<button>`` element not the value of the ``value`` attribute.

Our previous approach meant that IE6+7 users couldn't go back while creating a petition (we defaulted to 'next' in the controller) and couldn't advance at all when sponsoring or sigining (we didn't have the defaulting code and our ``Stages`` default to "stay" if they don't understand the move argument).

Hopefully using inputs instead of buttons won't mean we have problems styling things.